requirements.txt
flask
yt-dlp
apscheduler
python-dotenv
requests
google-genai
uvicorn
asgiref
whitenoise
Pillow
